pyproject.toml

[project]
name = "metric_at"
version = "0.1.0"
description = ""
authors = [
    {name = "Anna Chistyakova", email = "a.chistyakova@ispras.ru"},
]
dependencies = [
    "torch==2.0.1+cu118",
    "torchvision==0.15.2+cu118",
    "tensorboard>=2.13.0",
    "h5py>=3.9.0",
    "PyYAML>=6.0.1",
    "scipy>=1.9.3",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.2",
    "numpy>=1.25.2",
    "pandas>=2.0.3",
    "scikit-image>=0.21.0",
    "torchmetrics>=1.1.2",
    "pyiqa>=0.1.7",
    "scikit-learn>=1.3.1",
    "pytorch-msssim>=1.0.0",
    "seaborn>=0.13.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.277",
    "black>=23.7.0",
]

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.ruff]
line-length = 100

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/torch/"
name = "torch"

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/torchvision/"
name = "torchvision"

[tool.setuptools]
py-modules = ["metric_at"]

[project.scripts]
train = "metric_at.main:main"
